requirements.txt
alembic==1.4.2
aniso8601==8.0.0
appdirs==1.4.4
attrs==20.2.0
autopep8==1.5.4
Babel==2.8.0
blinker==1.4
certifi==2020.11.8
cfgv==3.2.0
Click==7.0
distlib==0.3.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.1
Flask-Babel==1.0.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Principal==0.4.0
flask-restx==0.2.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.3
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
identify==1.4.25
idna==2.9
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.24.1
nodeenv==1.4.0
passlib==1.7.2
pathtools==0.1.2
pre-commit==2.6.0
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
sentry-sdk==0.19.4
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.17
text-unidecode==1.3
toml==0.10.1
urllib3==1.26.2
virtualenv==20.0.30
visitor==0.1.3
wakatime==13.0.7
watchdog==0.10.2
Werkzeug==1.0.0
WTForms==2.3.1
zipp==3.1.0
